Fetch message count

+++ src/types.rs
@@ -3,7 +3,7 @@ pub struct ListedTopic {
     pub name: String,
     pub partitions: i32,
     pub replication_factor: i32,
-    pub message_count: i32,
+    pub message_count: i64,
     pub size: i64,
 }
